prefix_trie::joint

Module map

Module that defines the JointPrefixMap

Structs

CoverKeys

An iterator that yields all keys (prefixes) in a JointPrefixMap that covers a given prefix (including the prefix itself if preseint). See crate::joint::JointPrefixMap::cover_keys for how to create this iterator.

CoverValues

An iterator that yields all values in a JointPrefixMap that covers a given prefix (including the prefix itself if preseint). See crate::joint::JointPrefixMap::cover_values for how to create this iterator.

CoveringDifference

An iterator over the covering difference of two JointPrefixMaps, i.e., prefixes of self that are not covered by other. The iterator yields first prefixes of P1, followed by those of P2.

Difference

An iterator over the difference of two JointPrefixMaps, i.e., prefixes that are in self but not in other. The iterator yields first prefixes of P1, followed by those of P2.

DifferenceItem

An item of the Difference iterator.

Intersection

An iterator over the intersection of two JointPrefixMaps. The iterator yields first prefixes of P1, followed by those of P2.

IntoIter

An iterator over all owned entries of a JointPrefixMap in lexicographic order.

IntoKeys

An iterator over all prefixes of a JointPrefixMap in lexicographic order.

IntoValues

An iterator over all values of a JointPrefixMap in lexicographic order of their associated prefix.

Iter

An iterator over all entries of a JointPrefixMap in lexicographic order.

IterMut

A mutable iterator over a JointPrefixMap. This iterator yields elements in lexicographic order of their associated prefix.

JointPrefixMap

A Joint prefix map, implemented as two separate prefix trees.

Keys

An iterator over all prefixes of a JointPrefixMap in lexicographic order.

Union

An iterator over the union of two JointPrefixMaps. The iterator yields first prefixes of P1, followed by those of P2.

Values

An iterator over all values of a JointPrefixMap in lexicographic order of their associated prefixes.

ValuesMut

A mutable iterator over values of JointPrefixMap. This iterator yields elements in lexicographic order.

Enums

Cover

An iterator that yields all items in a JointPrefixMap that covers a given prefix (including the prefix itself if preseint). See crate::joint::JointPrefixMap::cover for how to create this iterator.

Entry

A mutable view into a single entry in a map, which may either be vacant or occupied.

OccupiedEntry

A mutable view into an occupied entry. An occupied entry represents a node that is already present on the tree.

UnionItem

An item of the Union iterator (over a JointPrefixMap).

VacantEntry

A mutable view into a missing entry. The information within this structure describes a path towards that missing node, and how to insert it.