prefix_trie::joint

Module set

JointPrefixSet, that is implemened as a simple binary tree, based on the super::JointPrefixMap.

Structs

CoveringDifference

An iterator over the covering difference of two JointPrefixSets. The iterator yields prefixes that are in self, but not covered by any prefix in other. The iterator yields first prefixes of P1, followed by those of P2.

Difference

An iterator over the difference of two JointPrefixSets. The iterator yields prefixes that are in self, but ont in other. The iterator also yields the longest prefix match in other (if it exists). The iterator yields first prefixes of P1, followed by those of P2.

Intersection

An iterator over the intersection of two JointPrefixSets. The iterator yields first prefixes of P1, followed by those of P2.

IntoIter

A consuming iterator over all entries of a JointPrefixSet in lexicographic order.

Iter

An iterator over all entries of a JointPrefixSet in lexicographic order.

JointPrefixSet

Set of prefixes, organized in a tree. This strucutre gives efficient access to the longest prefix in the set that contains another prefix.

Union

An iterator over the union of two JointPrefixSets. The iterator yields first prefixes of P1, followed by those of P2.