prefix_trie::joint

Trait JointPrefix

pub trait JointPrefix {
    type P1: Prefix + Clone;
    type P2: Prefix + Clone;

    // Required methods
    fn p1_or_p2(self) -> Either<Self::P1, Self::P2>;
    fn p1_or_p2_ref(&self) -> Either<&Self::P1, &Self::P2>;
    fn from_p1(p: &Self::P1) -> Self;
    fn from_p2(p: &Self::P2) -> Self;
}

Trait that defines a joint prefix, for instance, one that describes either a IPv4 or IPv6 address.

The joint prefix is implemented for ipnet::IpNet, [ipnetwork::IpNetwork], and [cidr::IpCidr]. It is also implemented for either::Either, for which the Left and Right types both implement crate::Prefix.

Required Associated Types

type P1: Prefix + Clone

The first prefix type, e.g., IPv4

type P2: Prefix + Clone

The first prefix type, e.g., IPv6

Required Methods

fn p1_or_p2(self) -> Either<Self::P1, Self::P2>

Get either Left(P1) or Right(P2).

fn p1_or_p2_ref(&self) -> Either<&Self::P1, &Self::P2>

Get either Left(P1) or Right(P2), as a reference.

fn from_p1(p: &Self::P1) -> Self

Construct a prefix from a reference to the first variant.

fn from_p2(p: &Self::P2) -> Self

Construct a prefix from a reference to the second variant.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.