prefix_trie

Trait Prefix

pub trait Prefix: Sized + Debug {
    type R: Unsigned + PrimInt + Zero + CheckedShr;

    // Required methods
    fn repr(&self) -> Self::R;
    fn prefix_len(&self) -> u8;
    fn from_repr_len(repr: Self::R, len: u8) -> Self;

    // Provided methods
    fn mask(&self) -> Self::R { ... }
    fn zero() -> Self { ... }
    fn longest_common_prefix(&self, other: &Self) -> Self { ... }
    fn contains(&self, other: &Self) -> bool { ... }
    fn is_bit_set(&self, bit: u8) -> bool { ... }
    fn eq(&self, other: &Self) -> bool { ... }
}

Trait for defining prefixes.

Required Associated Types

type R: Unsigned + PrimInt + Zero + CheckedShr

How can the prefix be represented. This must be one of u8, u16, u32, u64, or u128.

Required Methods

fn repr(&self) -> Self::R

Get raw representation of the address, ignoring the prefix length. This function must return the representation with the mask already applied.

fn prefix_len(&self) -> u8

Prefix length

fn from_repr_len(repr: Self::R, len: u8) -> Self

Create a new prefix from the representation and the prefix pength.

Provided Methods

fn mask(&self) -> Self::R

mask self.repr() using self.len(). If you can guarantee that repr is already masked, them simply re-implement this function for your type.

fn zero() -> Self

Create a prefix that matches everything

fn longest_common_prefix(&self, other: &Self) -> Self

longest common prefix

fn contains(&self, other: &Self) -> bool

Check if self contains other in its prefix range. This function also returns True if self is identical to other.

fn is_bit_set(&self, bit: u8) -> bool

Check if a specific bit is set (counted from the left, where 0 is the first bit from the left).

fn eq(&self, other: &Self) -> bool

Compare two prefixes together

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.