prefix_trie

Module trieview

A TrieView (or a TrieViewMut) is a pointer to a specific element in a PrefixTrie, representing the sub-tree rooted at that node.

This module allows you to perform Set operations (union, intersection, difference) on PrefixMaps and PrefixSets, optionally of only a trie-view.

Structs

CoveringDifference

An iterator over the covering difference of two TrieViews. See TrieView::covering_difference for more information and an example.

CoveringDifferenceMut

An iterator over the covering difference of two TrieViews. See TrieViewMut::covering_difference_mut for more information and an example.

Difference

An iterator over the difference of two TrieViews. See TrieView::difference for more information and an example.

DifferenceItem

An item of the Difference iterator.

DifferenceMut

An iterator over the difference of two TrieViews. See TrieViewMut::difference_mut for more information and an example.

DifferenceMutItem

An item of the DifferenceMut iterator.

Intersection

an iterator over the intersection of two crate::PrefixSets in lexicographic order.

IntersectionMut

an iterator over the intersection of two crate::PrefixSets in lexicographic order, yielding mutable references to all elements.

TrieView

A subtree of a prefix-trie rooted at a specific node.

TrieViewMut

A mutable view of a prefix-trie rooted at a specific node.

Union

An iterator over the union of two TrieViews that always yields either the exact value or the longest prefix match of both of them.

UnionMut

An iterator over the union of two TrieViews that always yields either the exact value or the longest prefix match of both of them.

Enums

UnionItem

An item of the Union iterator.

Traits

AsView

A trait for creating a TrieView of self.

AsViewMut

A trait for creating a TrieViewMut of self.