prefix_trie::trieview

Struct TrieView

pub struct TrieView<'a, P, T> { /* private fields */ }

A subtree of a prefix-trie rooted at a specific node.

The view can point to one of three possible things:

A node in the tree that is actually present in the map,

A branching node that does not exist in the map, but is needed for the tree structure (or that was deleted using the function remove_keep_tree)

A virtual node that does not exist as a node in the tree. This is only the case if you call TrieView::find or AsView::view_at with a node that is not present in the tree, but that contains elements present in the tree. Virtual nodes are treated as if they are actually present in the tree as branching nodes.

Implementations

impl<'a, P, L> TrieView<'a, P, L>
where P: Prefix,

pub fn difference<R>( &self, other: &'a impl AsView<P = P, T = R>, ) -> Difference<'a, P, L, R>

Iterate over the all elements in self that are not present in other. Each item will return a reference to the prefix and value in self, as well as the longest prefix match of other.

Warning: The iterator will only yield elements of the given views. If other points to a branching or a virtual node, then the longest prefix match returned may be None, even though it exists in the original map referenced by other.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map_a: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
    (net!("192.168.0.0/24"), 3),
    (net!("192.168.2.0/23"), 4),
]);
let mut map_b: PrefixMap<ipnet::Ipv4Net, &'static str> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), "a"),
    (net!("192.168.0.0/22"), "b"),
    (net!("192.168.0.0/23"), "c"),
    (net!("192.168.2.0/24"), "d"),
]);
let sub_a = map_a.view_at(net!("192.168.0.0/22")).unwrap();
let sub_b = map_b.view_at(net!("192.168.0.0/22")).unwrap();
assert_eq!(
    sub_a.difference(&sub_b).collect::<Vec<_>>(),
    vec![
        DifferenceItem { prefix: &net!("192.168.0.0/24"), value: &3, right: Some((&net!("192.168.0.0/23"), &"c"))},
        DifferenceItem { prefix: &net!("192.168.2.0/23"), value: &4, right: Some((&net!("192.168.0.0/22"), &"b"))},
    ]
);

pub fn covering_difference<R>( &self, other: &'a impl AsView<P = P, T = R>, ) -> CoveringDifference<'a, P, L, R>

Iterate over all elements in self that are not not covered in other. In other words, only items of self are yielded for which there does not exist a prefix in other that covers it.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map_a: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/22"), 1),
    (net!("192.168.0.0/24"), 2),
    (net!("192.168.2.0/23"), 3),
]);
let mut set_b: PrefixSet<ipnet::Ipv4Net> = PrefixSet::from_iter([
    net!("192.168.0.0/23"),
]);
assert_eq!(
    map_a.view().covering_difference(&set_b).collect::<Vec<_>>(),
    vec![(&net!("192.168.0.0/22"), &1), (&net!("192.168.2.0/23"), &3)],
);

impl<'a, P, L> TrieView<'a, P, L>
where P: Prefix,

pub fn intersection<R>( &self, other: &'a impl AsView<P = P, T = R>, ) -> Intersection<'a, P, L, R>

Iterate over the union of both Views. Each element will yield a reference to the prefix and the value stored in self and other (if the prefix is in both views).

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map_a: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
    (net!("192.168.0.0/24"), 3),
    (net!("192.168.2.0/23"), 4),
]);
let mut map_b: PrefixMap<ipnet::Ipv4Net, &'static str> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), "a"),
    (net!("192.168.0.0/22"), "b"),
    (net!("192.168.0.0/23"), "c"),
    (net!("192.168.0.0/24"), "d"),
    (net!("192.168.2.0/24"), "e"),
]);
let sub_a = map_a.view_at(net!("192.168.0.0/22")).unwrap();
let sub_b = map_b.view_at(net!("192.168.0.0/22")).unwrap();
assert_eq!(
    sub_a.intersection(&sub_b).collect::<Vec<_>>(),
    vec![
        (&net!("192.168.0.0/22"), &2, &"b"),
        (&net!("192.168.0.0/24"), &3, &"d"),
    ]
);

impl<'a, P, L> TrieView<'a, P, L>
where P: Prefix,

pub fn union<R>( &self, other: &'a impl AsView<P = P, T = R>, ) -> Union<'a, P, L, R>

Iterate over the union of two views. If a prefix is present in both trees, the iterator will yield both elements. Otherwise, the iterator will yield the element of one TrieView together with the longest prefix match in the other TrieView. Elements are of type UnionItem.

Warning: The iterator will only yield elements of the given TrieViews. If either of the two TrieViews is pointing to a branching or a virtual node, then the longest prefix match returned may be None, even though it exists in the larger tree.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map_a: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
    (net!("192.168.0.0/24"), 3),
]);
let mut map_b: PrefixMap<ipnet::Ipv4Net, &'static str> = PrefixMap::from_iter([
    (net!("192.168.0.0/22"), "a"),
    (net!("192.168.0.0/23"), "b"),
]);
assert_eq!(
    map_a.view().union(&map_b).collect::<Vec<_>>(),
    vec![
        UnionItem::Left{
            prefix: &net!("192.168.0.0/20"),
            left: &1,
            right: None,
        },
        UnionItem::Both{
            prefix: &net!("192.168.0.0/22"),
            left: &2,
            right: &"a",
        },
        UnionItem::Right{
            prefix: &net!("192.168.0.0/23"),
            left: Some((&net!("192.168.0.0/22"), &2)),
            right: &"b",
        },
        UnionItem::Left{
            prefix: &net!("192.168.0.0/24"),
            left: &3,
            right: Some((&net!("192.168.0.0/23"), &"b")),
        },
    ]
);

impl<'a, P, T> TrieView<'a, P, T>
where P: Prefix,

pub fn find(&self, prefix: P) -> Option<TrieView<'a, P, T>>

Find prefix, returning a new view that points to the first node that is contained within that prefix (or prefix itself). Only the current view is searched. If prefix is not present in the current view referenced by self (including any sub-prefix of prefix), the function returns None.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
    (net!("192.168.0.0/24"), 3),
    (net!("192.168.2.0/23"), 4),
    (net!("192.168.4.0/22"), 5),
]);
let sub = map.view();
assert_eq!(
    sub.find(net!("192.168.0.0/21")).unwrap().keys().collect::<Vec<_>>(),
    vec![
        &net!("192.168.0.0/22"),
        &net!("192.168.0.0/24"),
        &net!("192.168.2.0/23"),
        &net!("192.168.4.0/22"),
    ]
);
assert_eq!(
    sub.find(net!("192.168.0.0/22")).unwrap().keys().collect::<Vec<_>>(),
    vec![
        &net!("192.168.0.0/22"),
        &net!("192.168.0.0/24"),
        &net!("192.168.2.0/23"),
    ]
);

pub fn find_exact(&self, prefix: &P) -> Option<TrieView<'a, P, T>>

Find prefix, returning a new view that points to that node. Only the current view is searched. If this prefix is not present in the view pointed to by self, the function returns None.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
    (net!("192.168.0.0/24"), 3),
    (net!("192.168.2.0/23"), 4),
    (net!("192.168.4.0/22"), 5),
]);
let sub = map.view();
assert!(sub.find_exact(&net!("192.168.0.0/21")).is_none());
assert_eq!(
    sub.find_exact(&net!("192.168.0.0/22")).unwrap().keys().collect::<Vec<_>>(),
    vec![
        &net!("192.168.0.0/22"),
        &net!("192.168.0.0/24"),
        &net!("192.168.2.0/23"),
    ]
);

// If the prefix does not exist, the function returns `None`:
assert_eq!(map.view().find_exact(&net!("10.0.0.0/8")), None);

pub fn find_lpm(&self, prefix: &P) -> Option<TrieView<'a, P, T>>

Find the longest match of prefix, returning a new view that points to that node. Only the given view is searched. If the prefix is not present in the view pointed to by self, the function returns None.

Only views to nodes that are present in the map are returned, not to branching nodes.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
    (net!("192.168.0.0/24"), 3),
    (net!("192.168.2.0/23"), 4),
    (net!("192.168.4.0/22"), 5),
]);
let sub = map.view();
assert_eq!(
    sub.find_lpm(&net!("192.168.0.0/21")).unwrap().keys().collect::<Vec<_>>(),
    vec![
        &net!("192.168.0.0/20"),
        &net!("192.168.0.0/22"),
        &net!("192.168.0.0/24"),
        &net!("192.168.2.0/23"),
        &net!("192.168.4.0/22"),
    ]
);
assert_eq!(
    sub.find_lpm(&net!("192.168.0.0/22")).unwrap().keys().collect::<Vec<_>>(),
    vec![
        &net!("192.168.0.0/22"),
        &net!("192.168.0.0/24"),
        &net!("192.168.2.0/23"),
    ]
);

pub fn left(&self) -> Option<Self>

Get the left branch at the current view. The right branch contains all prefix that are contained within self.prefix(), and for which the next bit is set to 0.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let map: PrefixSet<ipnet::Ipv4Net> = PrefixSet::from_iter([
    net!("1.0.0.0/8"),
    net!("1.0.0.0/16"),
    net!("1.0.128.0/17"),
    net!("1.1.0.0/16"),
]);

let view = map.view_at(net!("1.0.0.0/8")).unwrap();
assert_eq!(view.prefix(), &net!("1.0.0.0/8"));

let view = view.left().unwrap();
assert_eq!(view.prefix(), &net!("1.0.0.0/15"));

let view = view.left().unwrap();
assert_eq!(view.prefix(), &net!("1.0.0.0/16"));

assert!(view.left().is_none());

pub fn right(&self) -> Option<Self>

Get the right branch at the current view. The right branch contains all prefix that are contained within self.prefix(), and for which the next bit is set to 1.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let map: PrefixSet<ipnet::Ipv4Net> = PrefixSet::from_iter([
    net!("1.0.0.0/8"),
    net!("1.0.0.0/16"),
    net!("1.1.0.0/16"),
    net!("1.1.0.0/24"),
]);

let view = map.view_at(net!("1.0.0.0/8")).unwrap();
assert_eq!(view.prefix(), &net!("1.0.0.0/8"));

assert!(view.right().is_none());
let view = view.left().unwrap();
assert_eq!(view.prefix(), &net!("1.0.0.0/15"));

let view = view.right().unwrap();
assert_eq!(view.prefix(), &net!("1.1.0.0/16"));

assert!(view.right().is_none());

impl<'a, P, T> TrieView<'a, P, T>

pub fn iter(&self) -> Iter<'a, P, T>

Iterate over all elements in the given view (including the element itself), in lexicographic order.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
    (net!("192.168.0.0/24"), 3),
    (net!("192.168.2.0/23"), 4),
]);
let sub = map.view_at(net!("192.168.0.0/22")).unwrap();
assert_eq!(
    sub.iter().collect::<Vec<_>>(),
    vec![
        (&net!("192.168.0.0/22"), &2),
        (&net!("192.168.0.0/24"), &3),
        (&net!("192.168.2.0/23"), &4),
    ]
);

pub fn keys(&self) -> Keys<'a, P, T>

Iterate over all keys in the given view (including the element itself), in lexicographic order.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
    (net!("192.168.0.0/24"), 3),
    (net!("192.168.2.0/23"), 4),
]);
let sub = map.view_at(net!("192.168.0.0/22")).unwrap();
assert_eq!(
    sub.keys().collect::<Vec<_>>(),
    vec![&net!("192.168.0.0/22"), &net!("192.168.0.0/24"), &net!("192.168.2.0/23")]
);

pub fn values(&self) -> Values<'a, P, T>

Iterate over all values in the given view (including the element itself), in lexicographic order.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
    (net!("192.168.0.0/24"), 3),
    (net!("192.168.2.0/23"), 4),
]);
let sub = map.view_at(net!("192.168.0.0/22")).unwrap();
assert_eq!(sub.values().collect::<Vec<_>>(), vec![&2, &3, &4]);

pub fn prefix(&self) -> &P

Get a reference to the prefix that is currently pointed at. This prefix might not exist explicitly in the map/set, but may be used as a branching node (or when you call remove_keep_tree).

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
]);

assert_eq!(map.view_at(net!("192.168.0.0/20")).unwrap().prefix(), &net!("192.168.0.0/20"));
assert_eq!(map.view_at(net!("192.168.0.0/21")).unwrap().prefix(), &net!("192.168.0.0/21"));
assert_eq!(map.view_at(net!("192.168.0.0/22")).unwrap().prefix(), &net!("192.168.0.0/22"));

pub fn value(&self) -> Option<&'a T>

Get a reference to the value at the root of the current view. This function may return None if self is pointing at a branching node.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
]);

assert_eq!(map.view_at(net!("192.168.0.0/20")).unwrap().value(), Some(&1));
assert_eq!(map.view_at(net!("192.168.0.0/21")).unwrap().value(), None);
assert_eq!(map.view_at(net!("192.168.0.0/22")).unwrap().value(), Some(&2));

pub fn prefix_value(&self) -> Option<(&'a P, &'a T)>

Get a reference to both the prefix and the value. This function may return None if either self is pointing at a branching node.

macro_rules! net { ($x:literal) => {$x.parse::<ipnet::Ipv4Net>().unwrap()}; }

let mut map: PrefixMap<ipnet::Ipv4Net, usize> = PrefixMap::from_iter([
    (net!("192.168.0.0/20"), 1),
    (net!("192.168.0.0/22"), 2),
]);

assert_eq!(
    map.view_at(net!("192.168.0.0/20")).unwrap().prefix_value(),
    Some((&net!("192.168.0.0/20"), &1))
);
assert_eq!(map.view_at(net!("192.168.0.0/21")).unwrap().prefix_value(), None);