prefix_trie::trieview

Trait AsViewMut

pub trait AsViewMut<'a, P: Prefix, T>: Sized {
    // Required method
    fn view_mut(self) -> TrieViewMut<'a, P, T>;

    // Provided method
    fn view_mut_at(self, prefix: P) -> Option<TrieViewMut<'a, P, T>> { ... }
}

A trait for creating a TrieViewMut of self.

Required Methods

fn view_mut(self) -> TrieViewMut<'a, P, T>

Get a mutable view rooted at the origin (referencing the entire trie).

Provided Methods

fn view_mut_at(self, prefix: P) -> Option<TrieViewMut<'a, P, T>>

Get a mutable view rooted at the given prefix. If that prefix is not part of the trie, None is returned. Calling this function is identical to self.view().find(prefix).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.